# openapi_complete_v5.3.yaml
openapi: 3.1.0
info:
  title: DentalKiosk Shopware E-Commerce API
  description: 'Complete API for the DentalKiosk chatbot middleware used with Shopware
    6 and Dify.


    Features:

    - Product search & details

    - Cart action queue (add/remove/update/clear)

    - Orders lookup and reorder (by orderNumber)

    - Multimodal pipeline (image -> extract -> search -> top3 -> confirm)

    - Analytics ingestion + summary

    '
  version: 5.3.0
  contact:
    name: DentalKiosk Support
    url: https://dentalkiosk.de
servers:
- url: https://cbshop.commerce-crew.com
  description: Production Middleware Server
- url: http://localhost:3330
  description: Local Development
...
paths:
...
  /api/products/search:
    post:
      operationId: searchProducts
      summary: Search for products by keyword
      description: 'Search the product catalog using keywords. Returns product name,
        price, image, and availability.

        Use this when a customer asks about products, wants to find something, or
        mentions a product name.

        '
      tags:
      - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - query
              properties:
                query:
                  type: string
                  description: Search term (product name, brand, category, keyword,
                    or description)
                  example: Handschuhe
                limit:
                  type: integer
                  description: Maximum number of results (1-50)
                  default: 10
                  minimum: 1
                  maximum: 50
                page:
                  type: integer
                  description: Page number for pagination
                  default: 1
                  minimum: 1
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
...
  /api/checkout/payment-methods:
    get:
      operationId: getPaymentMethods
      summary: Get available payment methods
      description: List all available payment options for checkout.
      tags:
      - Checkout
      parameters:
      - $ref: '#/components/parameters/ContextTokenRequired'
      responses:
        '200':
          description: Payment methods
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  paymentMethods:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
...
  /api/products/resolve:
    post:
      operationId: resolveProduct
      summary: Resolve SKU/name to product UUID
      tags:
      - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  type: string
                limit:
                  type: integer
                  default: 5
              required:
              - identifier
      responses:
        '200':
          description: Resolved
        '404':
          description: Not found
  /api/products/from-image:
    post:
      operationId: productsFromImage
      summary: Image -> extract -> search -> top3 suggestions
      tags:
      - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                message:
                  type: string
                image:
                  type: string
                  description: Base64 data URL
              required:
              - image
      responses:
        '200':
          description: Suggestions
  /api/products/confirm:
    post:
      operationId: confirmProductSuggestion
      summary: Confirm one of top-3 suggestions and enqueue add-to-cart
      tags:
      - Products
      - Cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                selection_id:
                  type: string
                product_id:
                  type: string
                quantity:
                  type: integer
                  default: 1
                userId:
                  type: string
              required:
              - selection_id
              - product_id
      responses:
        '200':
          description: Queued
...
components:
  parameters:
...
    ContextTokenRequired:
      name: sw-context-token
      in: header
      required: true
      schema:
        type: string
      description: Customer session token (required for this operation)
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          description: Shopware product UUID
        name:
          type: string
          description: Product name
        description:
          type: string
          description: Product description (HTML stripped)
        productNumber:
          type: string
          description: SKU/Article number
        price:
          type: number
          description: Price in EUR
        currency:
          type: string
          default: EUR
        priceFormatted:
          type: string
          description: Formatted price with currency
          example: 29,99 €
        imageUrl:
          type: string
          nullable: true
          description: Main product image URL
        thumbnailUrl:
          type: string
          description: Thumbnail image URL
        url:
          type: string
          description: Full product page URL
        available:
          type: boolean
          description: Is the product available for purchase
        inStock:
          type: boolean
        stock:
          type: integer
          description: Current stock level
        manufacturer:
          type: string
          description: Manufacturer name
...
    ProductListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          example: 15 Produkte gefunden
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
          description: Total number of matching products
        page:
          type: integer
        limit:
          type: integer
...